requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
python-dateutil>=2.9.0